requirements.txt
aioredis==1.3.1
amqp==5.0.2
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.2.1
Automat==20.2.0
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
constantly==15.1.0
cryptography==3.4.6
daphne==3.0.1
dj-database-url==0.5.0
Django==3.1.5
django-on-heroku==1.1.2
django-phonenumber-field==5.0.0
django-redis==4.12.1
django-rest-framework==0.1.0
djangorestframework==3.12.2
gunicorn==20.0.4
h11==0.12.0
hiredis==1.1.0
httptools==0.1.1
hyperlink==21.0.0
idna==2.10
importlib-metadata==3.7.0
incremental==17.5.0
jsonfield==3.1.0
kombu==5.0.2
msgpack==1.0.2
msgpack-python==0.5.6
phonenumbers==8.12.16
prompt-toolkit==3.0.16
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pypng==0.0.20
PyQRCode==1.2.1
python-dateutil==1.5
python-decouple==3.4
pytz==2021.1
redis==3.5.3
requests==2.25.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.7.4.3
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
zipp==3.4.0
zope.interface==5.2.0
